Refactored age selector handler and template to be generic for any range dimension data
