+++ assets/templates.go
@@ -187,7 +187,7 @@ func templatesDatasetFilterAgeSelectorListTmpl() (*asset, error) {
 	return a, nil
 }
 
-var _templatesDatasetFilterAgeSelectorRangeTmpl = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x58\x4d\x6f\xe3\x36\x13\x3e\x7b\x7f\xc5\x40\xc8\x71\x69\xd9\xd9\xbc\xd9\x77\x0b\xd9\x40\xd0\xed\x02\x05\xd2\x2d\x90\xa0\x05\x7a\x0a\x28\x71\x2c\xb3\xe1\x87\x4a\x52\x76\x52\xc3\xff\xbd\xa0\x28\xd9\xa2\x64\xa7\xc9\x6d\x0b\xac\x2f\x26\xe7\x8b\xc3\x99\x67\x46\x24\x33\xc6\x37\x50\x08\x6a\xed\x22\xa1\xec\xcf\xda\x3a\xb2\xd2\xca\x11\xcb\xff\x46\x42\xe6\xd7\xc9\xf2\xdd\xa4\x2f\x53\xd1\x12\x09\x57\xce\x68\xcf\x89\x58\x5b\x43\xab\x0a\x4d\x43\x8f\x18\x85\x16\xc4\x33\x03\x67\xc8\x6a\xa9\x93\xdd\x0e\x1c\xca\x4a\x50\x87\x90\x54\xd4\x38\x4e\x85\x4d\x73\x83\x94\x15\xa6\x96\x79\x02\x53\xd8\xef\x83\xec\xc0\x04\xf8\x15\x88\x64\xe4\xea\x63\x18\x8a\x92\x7c\xf8\xd4\x19\x9e\x64\xeb\xf9\xd8\xfd\x87\x07\xc7\x9d\x40\x68\x76\xbb\x45\x5e\xae\x1d\xf9\x38\x9b\x25\xcb\x9b\x12\xb3\x74\x3d\x6f\x95\xb3\x94\xf1\x4d\xeb\xf7\x61\x78\x18\x75\x83\xf6\x1f\x60\x14\xaa\x42\x2b\x87\xca\x41\xad\x18\x1a\xc1\x15\x12\x2a\x04\x11\x5c\x3d\x5a\xef\x5f\xac\xd1\x8b\x20\x34\xbf\x17\xa2\x78\x2e\x8c\xaf\x89\xcd\xff\x8e\xb1\x39\x2b\x7e\xf9\xe9\x20\x7e\xd9\x13\x9f\x64\x0c\x1d\xe5\xc2\x82\xd1\x02\x17\x49\x69\x74\x5d\x25\x9d\x01\x49\x4d\xc9\x15\xc9\xb5\x73\x5a\x12\x72\xd5\xd3\x03\xc8\x6c\x2d\x25\x35\xcf\xcb\xcc\x56\x54\x75\x2a\x2d\x31\x59\xde\x22\x35\x0a\xa4\x36\x08\x34\xd7\xb5\x03\x43\x95\x4f\x85\x17\x5e\x66\x69\xa7\xdc\x37\x18\x05\x5b\xa1\x88\x96\x9b\x64\xeb\x0f\x67\xfc\x9a\xc1\x36\xe4\xfa\x5e\x4b\x84\x06\x07\x59\xba\xfe\xd0\xd7\x06\xc8\xaa\x81\xb6\xd3\x15\x21\xb3\x64\x79\xab\x0d\x4a\xe0\x95\xad\x25\x30\x2d\xb4\x01\xcb\x1d\x50\x89\xee\x3d\x14\x5a\x59\x2c\x1c\xba\xda\x00\x65\xbc\xe2\xb6\xe0\xaa\x04\x14\xdc\x4d\xe1\x17\x5a\x1b\x6e\x61\xc3\x1d\x45\x30\x6b\xad\x8a\xda\x02\xe3\x54\x4e\xe1\x77\xbe\xa1\xb2\xb6\xa0\xb4\x02\x49\xad\xa5\x20\x78\x8e\x46\x4f\xe1\x6b\x2d\x04\x6d\xe9\x4f\xdc\xcb\x58\x94\x53\xf8\x52\xdb\x02\x81\x33\xd8\xa0\x75\x3c\xaf\x45\x2d\x3b\x95\x2c\xad\xa2\x28\x1d\x11\xdc\xa2\x38\x24\xf0\x88\x80\x48\x20\x5b\x69\x23\xbb\x9d\x37\xe3\x53\x6d\x01\x0a\x81\xd4\x10\x81\x2b\x97\x00\x2d\x1c\xd7\x6a\x91\xa4\x2b\x2e\x1c\x1a\x9b\xee\x76\xd3\x2f\xcd\xf0\xe7\xcf\xfb\x7d\xca\xb8\x44\x65\xb9\x56\x36\x82\xd1\xdb\x60\x17\x2b\xac\xb1\x78\xcc\xf5\x13\xbc\x80\xb7\xc9\x24\xe3\xaa\xaa\x1d\xb8\xe7\x0a\x8f\x2a\xc9\xd0\xc6\xc3\x43\x23\x96\x00\x67\x8b\xe4\x46\x08\xa0\x25\xda\x04\x14\x95\xd8\x9f\x6f\xa8\xa8\x23\x42\xa3\x8f\x2c\x5a\x50\xd0\x1c\xc5\x78\x81\x86\x9c\xc0\x4a\x9b\x9e\x81\xbe\xe2\xa4\xa3\xf6\x8d\xa5\x8d\x5a\x3f\x02\x71\x22\xa3\x88\xbc\x1c\x88\x7e\xbd\xe5\x42\x17\x8f\xc3\xc0\x5d\x26\x71\x51\xf6\x73\xfd\x31\x59\xde\x30\xe6\xdd\x8b\x2a\x12\xb4\x81\x8c\xc2\xda\xe0\x6a\x91\xec\x76\xd3\xcf\xd4\xd1\xe9\x0d\x63\x5f\x8c\x96\xb7\xdc\xba\xe9\x6f\x77\xb7\xfb\x7d\xb2\xa4\x8c\xc1\xca\x68\x8f\x4e\xeb\xb2\x94\x2e\x5b\xfd\x77\x7d\x07\x7b\x3b\xc9\xb5\x61\x68\x0e\x9e\x95\x54\x08\x34\xcf\xc4\x4a\x38\xc3\x91\xec\x1c\x47\x94\x01\xa6\x2b\xfe\x14\x87\x7b\xb7\xbb\xd8\x50\x01\x3f\x2c\x20\xb8\x1d\xc0\x6a\x6f\xa4\xae\x95\xeb\xbe\x31\x07\x51\x81\xca\x8b\xda\x3a\x77\x86\x16\x0e\x1a\xdd\xf9\x50\x2c\x04\x07\x2e\xf8\x7b\xb8\xd8\x78\x79\xa1\x75\x05\xb3\x46\x3a\x96\x3d\x07\xfd\xf9\xe5\x01\xfa\xf3\xcb\x13\x19\xea\xdb\x98\xec\x76\x7c\x05\xf8\x17\x5c\x70\x68\x1c\x8c\x97\x88\xd6\x58\x73\x86\xde\x2a\x84\x81\x64\x09\x50\xc3\x29\x59\x73\xc6\x50\x2d\x92\x15\x15\x16\x47\xe6\x51\xb1\x81\xcd\x01\xbe\x4f\x02\x69\xde\xc2\xfc\x59\xd7\xaa\x44\xeb\x06\x66\x27\x93\x3f\x5a\xc6\xc0\xf2\x08\xec\xaf\xd8\xe2\xa0\x1c\xce\x78\x7d\x02\x5b\x4d\x23\x1f\x03\x2b\x22\x1f\x51\x15\x91\x45\x09\x15\x65\x8c\xab\x32\xd0\x87\x69\x89\x96\xb3\x28\xb0\x70\x84\x8a\x71\x18\xb2\xc0\x8b\x25\x61\xcb\x99\x5b\x13\xeb\xe3\x38\x6b\x27\x92\xf5\x26\xa2\xf4\x93\xd0\xaa\xce\x46\x18\x20\xd3\x95\x6f\xc8\x5d\xcf\x0a\xc6\x93\x65\xf8\xcf\xd2\xc0\x1d\xaa\x9d\x85\xf0\x45\x5b\xda\x25\xde\x79\xf6\xf4\xde\x51\xe3\xbe\xd6\x72\xc4\xf8\x49\x31\x4f\x1e\xa2\x66\x32\x19\xf8\xe3\xab\xcf\x77\x86\xf0\xff\x92\x3f\xa8\xd8\x6b\xcc\x9d\xf7\x04\xa8\x62\xa0\x37\xfe\x48\xf5\x2a\xb1\x73\xce\x64\x69\x08\xde\x30\xd9\x63\x08\x8e\x49\x2d\xe5\x5f\x3b\x80\x2f\xfb\xf9\xb1\x19\xcc\xbf\xc1\x0e\xf0\x86\x06\xa0\x05\x3b\x01\xce\x5f\x1b\xea\x20\x60\xdf\x4b\xff\x8d\xa5\x7f\x3a\xb6\xa3\xc2\xf8\x5e\xf7\xff\x95\xba\xbf\x3e\x94\xfd\x35\x8c\xa1\x0f\xbb\x1d\x0a\x8b\xfb\x7d\xef\x1e\xe2\xb1\x71\x05\x3d\x82\xc7\xc7\x15\xb4\x55\x10\x03\x75\xdc\x49\xde\x08\x7b\x87\x4f\x8e\x98\xe6\x86\x3c\x9c\x7a\x25\xdf\x06\x06\x60\x8c\x8e\xde\xb6\xce\x25\x77\x87\x83\x77\xee\x14\xe4\x4e\x91\xe6\x12\xdc\xbb\x17\x37\xd3\xd8\xf1\xe4\x98\xeb\x61\x50\xda\x23\xa9\xed\x62\x73\x87\x52\x6f\xb0\xdd\x7f\x72\x2a\x11\x11\xc4\x86\xcd\x62\x24\x73\xe2\x70\xed\x74\x35\xdc\xe6\xf1\xfc\x7b\x71\x38\x00\x7f\xc5\x6d\x80\x59\x7b\x00\xf6\x9e\x2a\xdc\x76\x87\x67\xba\x7c\x61\xd9\x37\x1c\xf1\xff\xdf\xfa\x02\xd1\x6f\x74\x22\xbf\xa7\x1b\xbc\x51\xec\xce\xdf\x48\x55\xeb\xd2\x28\x13\x95\xe1\xfe\x6e\x1d\x26\x6e\xcd\x8b\xc7\x30\xdc\x72\x86\x61\x94\xf3\x32\x0c\x56\xda\x5f\x5a\x5b\x5f\x5a\x10\x5c\x0e\x1f\x51\x20\xba\x3f\x04\xc0\x84\xa7\x8b\xe0\x07\x38\x0d\xe1\xae\x08\xa1\xee\xec\x21\x2e\xa3\xed\xb4\xbe\x72\xd5\x80\x24\xfa\xe6\x34\x20\xb9\x8a\x17\x4b\x06\xfb\xff\x91\xaa\x02\x45\x97\x8b\x30\x7b\x6b\x12\xa8\xe5\x0c\xcf\x81\x23\x3a\xc1\x5f\x1d\x2f\xaf\xb3\xce\x4b\x7f\x47\x8e\xeb\xb5\x4b\xe1\x75\x24\xd2\x76\xf8\xfe\x3a\xf1\x13\xc7\x48\x3b\x7c\xcd\xad\x3c\x8f\xca\x6f\x10\x03\xb0\xaa\x85\x20\xcd\x47\xed\x15\x70\x38\xc4\xe1\x44\x0d\x9f\x7a\x2e\x0c\x66\x48\xe8\xda\x5c\xab\xfe\xa3\xe1\xa9\xe4\x66\x69\x9c\xdc\x2c\x5d\x69\x23\x97\xef\x8e\x10\x8c\x34\xc6\xaf\x81\x00\xe3\xe7\xc0\xf6\xef\x9f\x00\x00\x00\xff\xff\x27\xff\x2c\x14\x5b\x15\x00\x00")
+var _templatesDatasetFilterAgeSelectorRangeTmpl = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x58\x4d\x6f\xdb\x38\x13\x3e\xbb\xbf\x62\x20\xe4\x58\x59\x71\x9a\x37\x7d\xbb\x90\x0d\x18\xdb\x16\x28\x90\xed\x21\xc1\xee\x35\xa0\xc4\x91\xcc\x0d\x3f\xb4\x24\x65\x27\x6b\xf8\xbf\x2f\x28\x4a\xb6\x28\xd9\x6e\x82\x5d\xa0\x3d\xd4\x17\x51\xc3\xf9\xe2\xcc\x33\x23\x0f\x53\xca\xd6\x90\x73\x62\xcc\x3c\x22\xf4\xcf\xda\xd8\xb8\x50\xd2\xc6\x86\xfd\x8d\x71\x3c\xbb\x89\x16\x6f\x26\x7d\x9e\x8a\x94\x18\x33\x69\xb5\x72\x3b\xc1\xd6\x46\x93\xaa\x42\xdd\xd0\x83\x8d\x5c\xf1\xd8\x6d\xfa\x9d\xe1\x56\x4b\x9d\x6c\xb7\x60\x51\x54\x9c\x58\x84\xa8\x22\xda\x32\xc2\x4d\x92\x69\x24\x34\xd7\xb5\xc8\x22\x98\xc2\x6e\xe7\x79\x07\x2a\xc0\x59\x88\x05\x8d\xaf\xdf\xfb\x25\x2f\xe3\x77\x1f\x3a\xc5\x93\x74\x35\x1b\xbb\xff\xf0\x60\x99\xe5\x08\xcd\x69\x37\xc8\xca\x95\x8d\xdf\x5f\x5e\x46\x8b\x65\x89\x69\xb2\x9a\xb5\xc2\x69\x42\xd9\xba\xf5\x7b\xbf\xdc\xaf\xba\x45\xfb\x04\x18\x85\x2a\x57\xd2\xa2\xb4\x50\x4b\x8a\x9a\x33\x89\x31\xe1\x3c\xe6\x4c\x3e\x1a\xe7\x5f\x28\xd1\x8b\x20\x34\xbf\x33\x51\x3c\x15\xc6\x97\xc4\xe6\x7f\x87\xd8\x9c\x64\xbf\xfa\xb0\x67\xbf\xea\xb1\x4f\x52\x8a\x96\x30\x6e\x40\x2b\x8e\xf3\xa8\xd4\xaa\xae\xa2\x4e\x81\x20\xba\x64\x32\xce\x94\xb5\x4a\xc4\xf1\x75\x4f\x0e\x20\x35\xb5\x10\x44\x3f\x2f\x52\x53\x11\xd9\x89\xb4\xc4\x68\x71\x8b\x44\x4b\x10\x4a\x23\x90\x4c\xd5\x16\x34\x91\x2e\x15\x8e\x79\x91\x26\x9d\x70\x5f\x61\x10\x6c\x89\x3c\x30\x37\x49\x57\xef\x4e\xf8\x75\x09\x1b\x9f\xeb\x7b\x25\x10\x1a\x1c\xa4\xc9\xea\x5d\x5f\x1a\x20\xad\x06\xd2\x56\x55\x71\x7c\x19\x2d\x6e\x95\x46\x01\xac\x32\xb5\x00\xaa\xb8\xd2\x60\x98\x05\x22\xd0\xbe\x85\x5c\x49\x83\xb9\x45\x5b\x6b\x20\x94\x55\xcc\xe4\x4c\x96\x80\x9c\xd9\x29\xfc\x46\x6a\xcd\x0c\xac\x99\x25\x08\x7a\xa5\x64\x5e\x1b\xa0\x8c\x88\x29\xfc\xc1\xd6\x44\xd4\x06\xa4\x92\x20\x88\x31\x04\x38\xcb\x50\xab\x29\x7c\xad\x39\x27\x2d\xfd\x89\x39\x1e\x83\x62\x0a\x9f\x6b\x93\x23\x30\x0a\x6b\x34\x96\x65\x35\xaf\x45\x27\x92\x26\x55\x10\xa5\x03\x82\x5b\x14\xfb\x04\x1e\x10\x10\x30\xa4\x85\xd2\xa2\x3b\x79\xb3\x3e\xd6\x16\x20\xe7\x48\x74\xcc\xb1\xb0\x11\x90\xdc\x32\x25\xe7\x51\x52\x30\x6e\x51\x9b\x64\xbb\x9d\x7e\x6e\x96\x5f\x3e\xee\x76\x09\x65\x02\xa5\x61\x4a\x9a\x00\x46\xaf\x83\x5d\x28\xb0\xc2\xfc\x31\x53\x4f\x70\x06\x6f\x93\x49\xca\x64\x55\x5b\xb0\xcf\x15\x1e\x44\xa2\xa1\x8e\x87\x87\x86\x2d\x02\x46\xe7\xd1\x92\x73\x20\x25\x9a\x08\x24\x11\xd8\x7f\x5f\x13\x5e\x07\x84\x46\x1e\x69\x60\x90\x93\x0c\xf9\xd8\x40\x43\x8e\xa0\x50\xba\xa7\xa0\x2f\xe8\x7a\xdf\xc5\xf4\x23\xb1\x64\xba\xa4\x74\xc9\xf9\x17\x79\xe7\xc0\x3f\xbd\x6d\x14\x76\x5d\xaf\xcd\x56\xa3\xae\x1f\x99\x30\xc1\x41\xa4\xce\x07\xa8\x5f\x87\x19\x57\xf9\xe3\x30\xa0\x57\x51\x58\xac\x7d\x0c\xbc\x8f\x16\xdb\xed\xde\xe9\xcf\x5a\x89\x5b\x66\xac\x77\x79\xb7\x6b\xeb\x16\x94\x86\x94\xc0\x4a\x63\x31\x8f\x8e\xb1\xff\x7e\x77\xbb\xdb\x45\x0b\x42\x29\x14\x5a\x39\x0c\x1b\x9b\x26\x64\xd1\xca\xbf\xe9\xbb\xdb\x3b\x57\xa6\x34\x45\xbd\xf7\xb3\x24\x9c\xa3\x7e\x8e\x8d\x80\x13\x3b\x82\x9e\xda\xe1\xa5\x07\x73\xc1\x9e\x86\x49\xb9\x58\x13\x0e\xbf\xcc\xc1\xbb\xed\x21\x6d\x96\x42\xd5\xd2\x06\x39\x71\xac\x1c\xa5\x63\x35\x75\x66\x35\xc9\x2d\x34\xb2\xb3\x21\x9b\x6f\x6a\x70\xc1\xde\xc2\xc5\xda\xf1\x73\xa5\x2a\xb8\x6c\xb8\x43\xde\x53\x05\x32\xbb\xda\x17\xc8\xec\xea\x48\xbe\xfa\x3a\x26\xdb\x2d\x2b\x00\xff\x82\x0b\x06\x8d\x83\xa1\x89\xc0\xc6\x8a\x51\x74\x5a\xc1\x2f\x04\x8d\x80\x68\x46\xe2\x15\xa3\x14\xe5\x3c\x2a\x08\x37\x38\x52\x8f\x92\x0e\x74\x0e\xaa\xe0\x28\xac\x66\x6d\x31\x1c\x60\xdf\xc0\xbd\x59\xdd\x5b\xa2\x6d\x87\xfb\x81\xbd\x7e\xa1\x9c\x90\x18\xf8\x32\x2a\x96\x17\x04\x65\x50\x4e\x27\xce\x79\x04\x8d\xcd\x07\x62\x0c\xc5\x80\x7c\xc0\x61\x40\xe6\x25\x54\x84\x52\x26\x4b\x4f\x1f\x26\x32\x30\x67\x90\x63\x6e\x63\xc2\xed\x28\x3e\xa9\xdf\x0b\x39\x61\xc3\xa8\x5d\xc5\xc6\x45\xfe\xb2\x7d\x11\xb4\xf7\xc2\x4b\xf7\xe2\x5b\xe0\xeb\x73\x02\x90\xaa\xca\x7d\x01\xba\x26\xe9\xad\x46\x0b\xff\x4c\x13\xbf\x3b\x14\x3b\x59\x0d\xc7\xcd\x7f\xad\xc5\x78\xe7\x93\xa4\x8e\x3e\x4c\xfb\x64\x32\xf0\xc8\x95\xb2\xf3\xdc\x3f\xcf\x79\x84\x92\xbe\x44\xdd\x19\x57\x80\x48\x0a\x6a\xed\xfe\xc6\x9d\xe7\xdb\x6e\x81\x15\x63\x86\x65\x55\xa1\xa4\xf7\x56\xbb\x3f\x0e\xbb\xdd\x31\x1d\x23\x96\x06\x9e\xa7\x8e\x95\x26\x3e\x11\x43\x44\x8d\x71\x3e\x26\xb5\x94\x6f\x36\x26\xd7\x8d\x66\x87\x1e\x35\xfb\x01\x1b\xd3\xbf\xec\x4b\x9f\x24\x3d\x55\x01\xdf\x60\x1f\x84\xf8\x67\x47\xfa\x0f\x3a\xd2\x99\x6c\x8c\xaa\xf5\x67\x3b\xfa\x3e\xed\xe8\x3b\xf7\xa3\x9b\x7d\x3b\xba\x81\x71\x81\xc1\x76\x8b\xdc\xe0\x6e\xd7\x1b\xee\x1c\x02\xaf\xa1\x47\x70\x28\xbc\x86\xf6\x3c\x61\x39\x8c\x3b\xdc\x2b\x8b\xcb\xe2\x93\x8d\x75\x73\xed\x30\x7c\x75\x42\xae\x3d\x0d\x90\x1d\xcc\x33\xa6\xce\x04\xb3\xfb\x69\x26\xb3\x12\x32\x2b\xe3\xe6\x66\xa1\x77\xd9\xd0\xbc\x86\x8e\x47\x07\xd8\x0c\x83\xb2\xa4\xb4\x19\x4e\xba\xd8\xdc\xa1\x50\x6b\x6c\xcf\x1f\x1d\x4b\x44\x80\xd6\x61\x4b\x1a\xf1\x1c\x99\x4c\xac\xaa\x86\xc7\x3c\x8c\x0b\x87\x99\xe8\x2b\x6e\xfc\x44\xd4\xce\x0b\xce\x53\x89\x9b\xee\x8e\x80\x2c\xce\x98\x7d\xc5\x7c\xf4\xff\xd6\x17\x08\x7e\xa3\x01\xe6\x9e\xac\x71\x29\xe9\x9d\x1b\xf3\x65\xeb\xd2\x28\x13\x95\x66\x82\xe8\x67\xff\x62\x57\x2c\x7f\xf4\xcb\x0d\xa3\xe8\x57\x19\x2b\xfd\xa2\x50\x79\x6d\x3a\x40\xb5\x20\xb8\x1a\xde\x4c\x41\x30\x7c\x79\xc0\xf8\xfb\x20\xef\x07\x58\x05\x7e\x00\x07\x5f\x77\x66\x1f\x97\xd1\x71\x5a\x5f\x99\x6c\x40\x12\x7c\x0b\x1b\x90\x5c\x87\xc6\xa2\xc1\xf9\x7f\x25\x32\x47\xde\xe5\xc2\xbf\xbd\x36\x09\xc4\x30\x8a\xa7\xc0\x11\x0c\x3c\xd7\x87\x1b\x81\xcb\xce\x4b\x8e\x85\x0d\xeb\xb5\x4b\xe1\x4d\xc0\xd2\x7e\x47\xfa\x76\xc2\x7b\xa3\x91\xb4\xff\x97\x61\xc4\x69\x54\xfe\x80\x18\x80\xa2\xe6\x3c\x6e\x3e\x9d\x2f\x80\xc3\x3e\x0e\x47\x6a\xf8\xd8\x1d\xac\x57\x13\xfb\xae\xcd\x94\xec\xdf\xc4\x1e\x4b\x6e\x9a\x84\xc9\x4d\x93\x42\x69\xb1\x78\x73\x80\x60\x20\x31\xbe\x62\x05\x18\xdf\xb1\xb6\x8f\x7f\x02\x00\x00\xff\xff\x2b\x62\x3a\xb5\xb0\x16\x00\x00")
 
 func templatesDatasetFilterAgeSelectorRangeTmplBytes() ([]byte, error) {
 	return bindataRead(
@@ -202,12 +202,12 @@ func templatesDatasetFilterAgeSelectorRangeTmpl() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "templates/dataset-filter/age-selector-range.tmpl", size: 5467, mode: os.FileMode(420), modTime: time.Unix(1500462055, 0)}
+	info := bindataFileInfo{name: "templates/dataset-filter/age-selector-range.tmpl", size: 5808, mode: os.FileMode(420), modTime: time.Unix(1501500960, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
 
-var _templatesDatasetFilterFilterOverviewTmpl = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xac\x56\xcd\x6e\xe3\x36\x10\x3e\xa7\x4f\x31\x10\x7c\xd8\x05\x96\xfe\xab\x91\x74\x0b\xc5\xc0\x62\x83\x9e\x72\x28\x0a\xf4\x6c\x50\xe2\x58\x62\x4d\x91\x02\x49\xdb\x48\x09\xbe\xfb\x82\xa2\x24\xff\xc8\x4e\x1c\x24\xbc\x68\x38\x33\xfc\x38\xf3\xcd\x90\x54\xca\xf8\x0e\x72\x41\x8d\x79\x4c\x28\xfb\x6f\x6b\x2c\x59\x2b\x69\x89\xe1\xff\x23\x21\xb3\xfb\x64\xf9\xdb\xdd\xb1\x4f\x4d\x0b\x24\x5c\x5a\xad\x82\xe5\xc4\xb4\xd7\xb4\xae\x51\x37\xfa\x13\x43\xae\x04\x09\xc6\x68\x39\x37\xb5\xda\x3b\xe7\xc0\x62\x55\x0b\x6a\x11\x92\x9a\x6a\xcb\xa9\x30\x93\x4c\x23\x65\xb9\xde\x56\x59\x02\x63\xf0\x3e\xfa\x9e\x41\x40\xd8\x81\x54\x8c\x2c\x1e\xa2\x28\x0a\xf2\xfb\xf7\x0e\xf8\x2e\x2d\x67\xc3\xf0\x57\x2b\xcb\xad\x40\x68\xb2\xdd\x23\x2f\x4a\x4b\x1e\xa6\xd3\x64\xf9\x17\x17\x16\x35\xa8\xda\x72\x25\x4d\x3a\x29\x67\x2d\x4e\x3a\x61\x7c\xd7\xa6\xd0\x8b\xbd\xd4\x09\xed\x17\x60\xc0\x5a\xae\xa4\x45\x69\x61\x2b\x19\x6a\xc1\x25\x12\x2a\x04\x11\x5c\x6e\x4c\x12\x16\x84\x71\x99\x4f\xe8\xc7\x15\x5a\xe1\x64\x0c\x09\x86\xc1\xb8\x89\x41\xa8\xa8\x2e\xb8\x24\x99\xb2\x56\x55\x84\x2c\x2e\x42\x35\x70\x5b\xd1\xa1\x09\x6e\x2c\x21\x12\xb7\x56\x53\x01\xeb\x86\x4d\xa2\x76\xa8\x77\x1c\xf7\x57\x01\x00\x52\xc1\x3b\x88\x92\x33\x24\xc4\x54\xdd\xfe\x02\xd7\x96\x90\x29\xd4\x94\x31\x2e\x8b\x3e\x9e\x79\xaf\xb1\xaa\x0e\xd3\x4c\x69\x86\xba\xb7\x17\x54\x08\xd4\x2f\x01\xe9\x8a\xa5\x62\xd0\xd1\x08\x7b\xce\x6c\x19\x12\x8f\xbd\x73\x2d\xce\xd7\xd8\x9b\x2f\x7a\xf6\x66\xdf\x93\x65\x5a\xce\x3b\xaf\xe3\x43\xf5\xd0\xe5\xd5\x44\x3d\x4d\x96\x4f\xbc\x42\x69\xda\x76\x9b\x2f\xfb\x16\x7a\x7f\x00\xb3\x43\xf9\x66\x8b\xdb\x03\x88\x3d\x7f\xd3\xee\xe9\x44\xf0\xeb\x56\xe7\x40\x53\x59\x20\x8c\x9f\xa8\xa5\xe3\x43\x62\xde\xdf\x52\xf8\x4f\xaf\x77\xe3\x77\x54\xec\x77\xb6\xc1\x45\x18\x53\x7d\x52\x77\x18\xab\x95\x2c\x96\xce\x8d\x23\xff\xde\xa7\x93\x56\xf7\x59\x2d\x70\x7e\xfe\x56\xab\xa8\x30\xc9\xd2\xb9\x91\x40\x59\xd8\x12\xfe\x7c\x04\x81\x12\xc6\x3f\x18\x43\xf6\x93\x5a\x2c\x94\xe6\x68\xbc\x77\x6e\x94\xf7\xd3\xe0\x36\x74\xf9\xd2\xc3\x78\xff\xd5\xb9\x58\xfc\x11\xff\x06\xa3\x3c\x2c\x38\x5a\xef\x3d\x04\xbc\x80\xca\xd7\x20\x95\x85\x2f\x82\x1a\x0b\x23\x7e\xec\xf5\x15\xbc\xff\xe6\x1c\x4a\x16\x1c\x9b\xcf\x07\xb8\xb8\xef\xa9\xb8\x7f\xa3\x66\x0d\x0a\xed\x0f\x8b\x50\xd4\x12\x14\x84\xe8\xe6\x61\xa8\x18\x9c\xab\x44\x91\x40\xa9\x71\xfd\x98\x38\x37\x7e\xe6\x72\x33\xfe\xf7\x9f\x67\xef\x03\xaf\x71\xfa\x4c\x33\x14\x21\x7a\xfa\x46\xec\x1f\x3c\x6f\x28\x19\x5c\x3d\x5d\xe9\x64\x2b\x2e\x5e\xff\x97\x36\x1d\x28\x87\x8a\x23\x9e\xcf\xde\x86\x3f\x86\xcf\x50\x4f\x67\x66\x25\x64\x56\x12\x52\x6b\x5e\x51\xfd\x12\x27\xb6\xe4\xf9\x26\x8a\x7b\xce\x30\x4a\x19\x2f\xa2\xb0\x56\xf9\xd6\x74\xf7\x54\xa4\x9c\xcc\xcf\x5f\x6b\x38\xb9\xd5\x8e\x0a\xd2\xdc\x3e\x7f\x6b\x0c\x3d\xff\x43\xb2\x27\xb5\x97\x42\x51\xd6\xd5\xa8\xb5\x00\x95\x0c\x58\x6b\xbb\x50\xa8\x43\x06\x5c\x36\x0f\x76\x26\x54\xbe\x39\xb9\x3c\x17\xaf\x87\xf0\x93\xca\x1c\x45\xb7\x6d\x9c\x9d\x6d\x74\x42\xf2\xe1\xef\x21\x0a\xed\xe7\x57\x00\x00\x00\xff\xff\x02\x18\x17\xb7\xa6\x09\x00\x00")
+var _templatesDatasetFilterFilterOverviewTmpl = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xac\x56\xcd\x6e\xe3\x38\x0c\x3e\x77\x9f\x82\x30\x72\x68\x81\x2a\x7f\x1b\xb4\xdb\x85\x1b\xa0\x68\xb1\xa7\x1e\x16\x03\xcc\x39\x90\x2d\xc6\xd6\x44\x96\x0c\x59\x49\xd0\x11\xf4\xee\x03\x59\xb6\xf3\xe3\xa4\x4d\xd1\xea\x62\x8a\xa4\x3e\x91\x1f\x29\xc9\x31\xe3\x1b\x48\x05\xad\xaa\xc7\x88\xb2\x5f\xeb\xca\x90\xa5\x92\x86\x54\xfc\x37\x12\x32\xb9\x8b\xe6\x7f\x5d\xed\xfb\x94\x34\x43\xc2\xa5\xd1\xca\x5b\x0e\x4c\x5b\x4d\xcb\x12\x75\xad\x3f\x30\xa4\x4a\x10\x6f\x0c\x96\x63\x53\xa3\xbd\xb2\x16\x0c\x16\xa5\xa0\x06\x21\x2a\xa9\x36\x9c\x8a\x6a\x94\x68\xa4\x2c\xd5\xeb\x22\x89\x60\x08\xce\x05\xdf\x23\x08\xf0\x3b\x90\x82\x91\xd9\x7d\x10\x45\x46\xfe\x7e\x68\x81\xaf\xe2\x7c\xd2\x0f\x7f\xb1\x30\xdc\x08\x84\x3a\xdb\x2d\xf2\x2c\x37\xe4\x7e\x3c\x8e\xe6\xff\x71\x61\x50\x83\x2a\x0d\x57\xb2\x8a\x47\xf9\xa4\xc1\x89\x47\x8c\x6f\x9a\x14\x3a\xb1\x93\x5a\xa1\xf9\x02\xf4\x58\x4b\x95\x34\x28\x0d\xac\x25\x43\x2d\xb8\x44\x42\x85\x20\x82\xcb\x55\x15\xf9\x05\x7e\x9c\xe6\x13\xba\x71\x86\x56\x38\x18\x7d\x82\xa1\x37\x2e\x62\x10\x0a\xaa\x33\x2e\x49\xa2\x8c\x51\x05\x21\xb3\x93\x50\x35\xdc\x5a\xb4\x68\x82\x57\x86\x10\x89\x6b\xa3\xa9\x80\x65\xcd\x26\x51\x1b\xd4\x1b\x8e\xdb\xb3\x00\x00\xb1\xe0\x2d\x44\xce\x19\x12\x52\x15\xed\xfe\x02\x97\x86\x90\x31\x94\x94\x31\x2e\xb3\x2e\x9e\x69\xa7\x31\xaa\xf4\xd3\x44\x69\x86\xba\xb3\x67\x54\x08\xd4\x6f\x1e\xe9\x8c\xa5\x60\xd0\xd2\x08\x5b\xce\x4c\xee\x13\x0f\xbd\x73\x2e\xce\xf7\xd8\x9b\xec\xd8\x9b\xcc\xa2\x79\x9c\x4f\x5b\xaf\xfd\x43\x75\xdf\xe6\x55\x47\x3d\x8e\xe6\x2f\xbc\x40\x59\x35\xed\x36\x9d\x77\x2d\xf4\xf9\x00\xa6\xb3\x5d\x00\x0f\x97\x07\x10\x7a\xfe\xa2\xdd\xe3\x91\xe0\xe7\xad\xd6\x82\xa6\x32\x43\x18\xbe\x50\x43\x87\xbb\xc4\x9c\xbb\xa4\xf0\xdf\x5e\xef\xda\x6f\xaf\xd8\x9f\x6c\x83\x93\x30\x55\xf1\x4d\xdd\x51\x19\xad\x64\x36\xb7\x76\x18\xf8\x77\x2e\x1e\x35\xba\xef\x6a\x81\xe3\xf3\xb7\x58\x04\x45\x34\xb7\x76\x20\x50\x66\x26\x87\x7f\x1f\x41\xa0\x84\xe1\x13\x63\xc8\x9e\xa9\xc1\x4c\x69\x8e\x95\x73\xd6\x0e\xd2\x6e\xea\xdd\xfa\x2e\xd7\x1d\x8c\x73\x37\xd6\x86\xda\x0f\xf8\x2d\x0c\x52\xbf\x60\x6f\xbd\x73\xe0\xf1\x3c\x2a\x5f\x82\x54\x06\xae\x05\xad\x0c\x0c\xf8\xbe\xd7\x0d\x38\x77\x6b\x2d\x4a\xe6\x1d\xeb\xcf\x17\xa8\xb8\xeb\x98\xb8\xfb\xa0\x64\x35\x0a\xed\xce\x8a\x50\xd4\x10\x14\x84\xe8\xfa\x5d\x28\x18\x1c\xab\x44\x16\x41\xae\x71\xf9\x18\x59\x3b\x7c\xe5\x72\x35\xfc\xf9\xe3\xd5\x39\xcf\x6b\x98\xbe\xd2\x04\x85\x8f\x9e\x7e\x10\xfb\x17\x8f\x1b\x4a\x06\x67\x0f\x57\x3c\x5a\x8b\x93\xb7\xff\xa9\x4d\x7b\xca\xbe\x62\x8f\xe7\xa3\xa7\xe1\x9f\xfe\x2b\xd4\xd1\x99\x18\x09\x89\x91\x84\x94\x9a\x17\x54\xbf\x85\x89\xc9\x79\xba\x0a\xe2\x96\x33\x0c\x52\xc2\xb3\x20\x2c\x55\xba\xae\xda\x6b\x2a\x50\x4e\xa6\xc7\x8f\x35\x1c\x5c\x6a\x7b\x05\xa9\x2f\x9f\xff\x35\xfa\x96\x7f\x92\xec\x45\x6d\xa5\x50\x94\xb5\x35\x6a\x2c\x40\x25\x03\xd6\xd8\x4e\x14\x6a\x97\x01\x97\xf5\x7b\x9d\x08\x95\xae\x0e\xee\xce\xd9\xfb\x21\x3c\x53\x99\xa2\x68\xb7\x0d\xb3\xa3\x8d\x0e\x48\xde\xfd\x3c\x04\xa1\xf9\xfc\x09\x00\x00\xff\xff\x75\x61\xd5\xb0\xa5\x09\x00\x00")
 
 func templatesDatasetFilterFilterOverviewTmplBytes() ([]byte, error) {
 	return bindataRead(
@@ -222,7 +222,7 @@ func templatesDatasetFilterFilterOverviewTmpl() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "templates/dataset-filter/filter-overview.tmpl", size: 2470, mode: os.FileMode(420), modTime: time.Unix(1500462091, 0)}
+	info := bindataFileInfo{name: "templates/dataset-filter/filter-overview.tmpl", size: 2469, mode: os.FileMode(420), modTime: time.Unix(1500472047, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
