+++ handlers/dataset-filter/rangeSelector/handler.go
@@ -1,15 +1,15 @@
-package ageSelectorRange
+package rangeSelector
 
 import (
 	"net/http"
 
-	"github.com/ONSdigital/dp-frontend-models/model/dataset-filter/ageSelectorRange"
+	"github.com/ONSdigital/dp-frontend-models/model/dataset-filter/rangeSelector"
 	"github.com/ONSdigital/dp-frontend-renderer/render"
 )
 
-// Handler handles rendering of age selector range template
+// Handler handles rendering of range selector template
 func Handler(w http.ResponseWriter, req *http.Request) {
-	var page ageSelectorRange.Page
+	var page rangeSelector.Page
 
 	render.Handler(w, req, &page, &page.Page, "dataset-filter/age-selector-range", nil)
 }
